p_chara: fix signedness of m_keyId compare in CHandle::ChangeTexture
The texture-cache lookup compared it->m_keyId == (int)charaNo (signed cmpw),
but the target uses an unsigned cmplw. Cast m_keyId to unsigned long and compare
against the unsigned charaNo param, matching LoadModel's lookup idiom.
ChangeTexture 79.42->79.59%, unit 87.764->87.772%.

+++ src/p_chara.cpp
@@ -2247,7 +2247,7 @@ void CCharaPcs::CHandle::ChangeTexture(
     CLoadTexture* loadTexture;
     for (unsigned int i = 0; i < static_cast<unsigned int>(LoadTextureArray(&CharaPcs)->GetSize()); i++) {
         CLoadTexture* it = (*LoadTextureArray(&CharaPcs))[i];
-        if (reinterpret_cast<int>(it->m_keyTag) == charaKind && it->m_keyId == static_cast<int>(charaNo) &&
+        if (reinterpret_cast<int>(it->m_keyTag) == charaKind && static_cast<unsigned long>(it->m_keyId) == charaNo &&
             it->m_variantTag == reinterpret_cast<void*>(textureVariant)) {
             loadTexture = it;
             goto foundTexture;
